Refactor context handling and update package exports
