+++ node_package/src/ReactOnRails.web.ts
@@ -1,8 +1,5 @@
 import ReactOnRails from './ReactOnRails';
 import { setReactOnRails } from './context';
-import streamServerRenderedReactComponent from './streamServerRenderedReactComponent';
-
-ReactOnRails.streamServerRenderedReactComponent = streamServerRenderedReactComponent;
 
 setReactOnRails(ReactOnRails);
